implemented review changes
